website/requirements.txt
asgiref==3.3.1
certifi==2019.9.11
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
django-sslserver==0.22
gunicorn==20.0.4
idna==2.10
mysql-connector-python==8.0.22
oauthlib==3.1.0
protobuf==3.14.0
psycopg2==2.7.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
